fix: sql执行记录根据关键词搜索问题修复等

+++ server/internal/db/api/form/db_data_sync.go
@@ -26,5 +26,5 @@ type DataSyncTaskForm struct {
 
 type DataSyncTaskStatusForm struct {
 	Id     uint64 `binding:"required" json:"taskId"`
-	Status int    `json:"status"`
+	Status int8   `json:"status"`
 }
